requirements.txt
Django>=3.2,<4.0
djangorestframework
pika
celery
django-celery-beat
django-extensions
psycopg2-binary
requests